Update legend.vue
Adding the WHO limit value is relevant. All other limit values are reginally negociated. 25 is the current EU limit (which will change to 20, so I kept 20). There is no hour or day limit values; These are yearly men recommendations for WHO. When you are hiogher, you contribute to a higher year mean.
source:
https://environnement.brussels/etat-de-lenvironnement/rapport-2011-2014/air/qualite-de-lair-concentration-en-particules-tres-fines